bcg_deployment/requirements.txt
# Core dependencies for BCG Deployment Package
numpy>=1.20.0
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
Pillow>=8.0.0
scipy>=1.7.0
joblib>=1.0.0

# Optional: For development and testing
# pytest>=6.0.0
# pytest-cov>=2.12.0
